+++ generators/csharp/codegen/src/AsIs.ts
@@ -12,6 +12,8 @@ export enum AsIsFiles {
     DateTimeSerializer = "DateTimeSerializer.cs",
     EnumConverter = "EnumConverter.Template.cs",
     GrpcRequestOptions = "GrpcRequestOptions.Template.cs",
+    Headers = "Headers.Template.cs",
+    HeaderValue = "HeaderValue.Template.cs",
     HttpMethodExtensions = "HttpMethodExtensions.cs",
     JsonConfiguration = "JsonConfiguration.cs",
     OneOfSerializer = "OneOfSerializer.cs",